+++ server/routes/validate/validateNewLegalContact.ts
@@ -1,11 +1,5 @@
 import { AddNewLegalPinPhoneContactDetails } from 'express-session'
-import { validatePhoneNumber } from './validateTelephoneNumber'
-
-const errorMessages = {
-  phoneRequired: 'Enter the contact’s phone number',
-  invalidFormat: 'Enter a phone number in the correct format',
-  invalidNumber: 'You have entered an invalid number',
-}
+import { validateAndAssignError } from './validateTelephoneNumber'
 
 // eslint-disable-next-line import/prefer-default-export
 export const validateAddNewLegalContact = (form: AddNewLegalPinPhoneContactDetails) => {
@@ -27,25 +21,8 @@ export const validateAddNewLegalContact = (form: AddNewLegalPinPhoneContactDetai
     errors.relationship = { text: 'Select a relationship' }
   }
 
-  function validateAndAssignError(fieldName: 'telephone1' | 'telephone2', isRequired: boolean) {
-    const value = form[fieldName]?.trim() || ''
-    if (isRequired && !value) {
-      errors[fieldName] = { text: errorMessages.phoneRequired }
-      return
-    }
-
-    if (value) {
-      const result = validatePhoneNumber(value)
-      if (result === 'invalid_format') {
-        errors[fieldName] = { text: errorMessages.invalidFormat }
-      } else if (result === 'invalid_number') {
-        errors[fieldName] = { text: errorMessages.invalidNumber }
-      }
-    }
-  }
-
-  validateAndAssignError('telephone1', true)
-  validateAndAssignError('telephone2', false)
+  validateAndAssignError({ form, errors, fieldName: 'telephone1', isRequired: true })
+  validateAndAssignError({ form, errors, fieldName: 'telephone2', isRequired: false })
 
   return errors
 }
